+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 2018.12.10-23:45
+
++ Added `contains` function.
++ Added `delete` function.
++ Added `children` function to the `Node` class. to get ammount of children of the node.
+
 ## 2018.11.08-23:52
 
 + Added `BinarySearchTree`, `Node` and `BinarySearchTreeTest` java classes.